requirements.txt
Django==4.2.7
Pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
channels==4.0.0
channels-redis==4.1.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-widget-tweaks==1.5.0
